rename: Info.js to Info.tsx
