+++ benchexec/tablegenerator/react-table/scripts/dependencies.js
@@ -35,7 +35,7 @@ checker.init(
     // The following is the list of currently allowed licenses for bundled code.
     // We can add other free licenses, but must adjust the licences for
     // build/vendors.min* in ../../../.reuse/dep5, the linkification code in
-    // src/components/Info.js and pyproject.toml accordingly,
+    // src/components/Info.tsx and pyproject.toml accordingly,
     // and run "reuse download --all".
     onlyAllow: "BSD-3-Clause; CC-BY-4.0; ISC; MIT",
     customFormat: {
