+++ benchexec/tablegenerator/react-table/src/components/Overview.js
@@ -9,7 +9,7 @@ import React from "react";
 import { HashRouter as Router, Routes, Route, Link } from "react-router-dom";
 import Table from "./ReactTable.js";
 import Summary from "./Summary.js";
-import Info from "./Info.js";
+import Info from "./Info";
 import SelectColumn from "./SelectColumn.js";
 import ScatterPlot from "./ScatterPlot.js";
 import QuantilePlot from "./QuantilePlot.js";
